AMQP: flow stages - flow implementations splitted into separate files, minor refactors (after review)
