+++ amqp/src/main/scala/akka/stream/alpakka/amqp/javadsl/AmqpFlow.scala
@@ -15,12 +15,12 @@ import scala.compat.java8.FutureConverters._
 object AmqpFlow {
 
   /**
-   * Creates an [[AmqpFlow]] that accepts [[WriteMessage]] elements and emits [[WriteResult]].
+   * Creates an `AmqpFlow` that accepts `WriteMessage` elements and emits `WriteResult`.
    *
-   * This variant of [[AmqpFlow]] publishes messages in a fire-and-forget manner, hence all emitted [[WriteResult]]s
+   * This variant of `AmqpFlow` publishes messages in a fire-and-forget manner, hence all emitted `WriteResult`s
    * have `confirmed` flag set to `true`.
    *
-   * This stage materializes to a [[CompletionStage]] of [[Done]], which can be used to know when the Flow completes,
+   * This stage materializes to a `CompletionStage` of `Done`, which can be used to know when the Flow completes,
    * either normally or because of an amqp failure.
    */
   def create[T](
@@ -29,14 +29,14 @@ object AmqpFlow {
     akka.stream.alpakka.amqp.scaladsl.AmqpFlow(settings).mapMaterializedValue(f => f.toJava).asJava
 
   /**
-   * Creates an [[AmqpFlow]] that accepts [[WriteMessage]] elements and emits [[WriteResult]].
+   * Creates an `AmqpFlow` that accepts `WriteMessage` elements and emits `WriteResult`.
    *
-   * This variant of [[AmqpFlow]] waits for confirmation after every single message publication.
+   * This variant of `AmqpFlow` waits for confirmation after every single message publication.
    * It can be used to ensure order of messages accepted by queue at the cost of significantly lower publication
    * throughput. Please note that such strict ordering guarantee is rarely needed, and in most cases it's perfectly
    * sufficient to use `createWithAsyncConfirm` or `createWithAsyncUnorderedConfirm` for better performance.
    *
-   * This stage materializes to a [[CompletionStage]] of [[Done]], which can be used to know when the Flow completes,
+   * This stage materializes to a `CompletionStage` of `Done`, which can be used to know when the Flow completes,
    * either normally or because of an amqp failure.
    */
   def createWithConfirm[T](
@@ -52,9 +52,9 @@ object AmqpFlow {
       .asJava
 
   /**
-   * Creates an [[AmqpFlow]] that accepts [[WriteMessage]] elements and emits [[WriteResult]].
+   * Creates an `AmqpFlow` that accepts `WriteMessage` elements and emits `WriteResult`.
    *
-   * This variant of [[AmqpFlow]] asynchronously waits for message confirmations. Maximum number of messages
+   * This variant of `AmqpFlow` asynchronously waits for message confirmations. Maximum number of messages
    * simultaneously waiting for confirmation before signaling backpressure is configured with a
    * `bufferSize` parameter. Emitted results preserve the order of messages pulled from upstream - due to that
    * restriction this flow is expected to be slightly less effective than it's unordered counterpart.
@@ -62,7 +62,7 @@ object AmqpFlow {
    * In case of upstream failure/finish this stage attempts to process all buffered messages (waiting for
    * confirmation) before propagating failure/finish downstream.
    *
-   * This stage materializes to a [[CompletionStage]] of [[Done]], which can be used to know when the Flow completes,
+   * This stage materializes to a `CompletionStage` of `Done`, which can be used to know when the Flow completes,
    * either normally or because of an amqp failure.
    */
   def createWithAsyncConfirm[T](
@@ -80,9 +80,9 @@ object AmqpFlow {
       .asJava
 
   /**
-   * Creates an [[AmqpFlow]] that accepts [[WriteMessage]] elements and emits [[WriteResult]].
+   * Creates an `AmqpFlow` that accepts `WriteMessage` elements and emits `WriteResult`.
    *
-   * This variant of [[AmqpFlow]] asynchronously waits for message confirmations. Maximum number of messages
+   * This variant of `AmqpFlow` asynchronously waits for message confirmations. Maximum number of messages
    * simultaneously waiting for confirmation before signaling backpressure is configured with a
    * `bufferSize` parameter. Results are emitted downstream as soon as confirmation is received, meaning that
    * there is no ordering guarantee of any sort.
@@ -90,7 +90,7 @@ object AmqpFlow {
    * In case of upstream failure/finish this stage attempts to process all buffered messages (waiting for
    * confirmation) before propagating failure/finish downstream.
    *
-   * This stage materializes to a [[CompletionStage]] of [[Done]], which can be used to know when the Flow completes,
+   * This stage materializes to a `CompletionStage` of `Done`, which can be used to know when the Flow completes,
    * either normally or because of an amqp failure.
    */
   def createWithAsyncUnorderedConfirm[T](
